Split tests into meaningful modules
