+++ tests/test_wrappers.py
@@ -1,98 +1,11 @@
-"""Module for testing spiders."""
+"""Module for testing spider wrapper functionalities."""
 
 
-import json
 import logging
-import tempfile
 
 from pathlib import Path
 
 from finscraper.spiders import ISArticle, ILArticle
-
-
-# TODO: Implement utility test function that performs common Spider checks
-
-
-def test_ISArticle_with_category():
-    # Test scraping, no chromedriver
-    spider = ISArticle('ulkomaat').scrape(1)
-    df = spider.get()
-    assert len(df) >= 1
-    assert len(df.columns) == 8
-
-    # Test scraping with chromedriver
-    spider = ISArticle('ulkomaat', allow_chromedriver=True).scrape(1)
-    df = spider.get()
-    assert len(df) >= 1
-    assert len(df.columns) == 8
-
-    # Test continuing scraping
-    df2 = spider.scrape(1).get()
-    assert len(df2) >= len(df) + 1
-
-    # Save and load spider
-    jobdir = spider.save()
-    spider = ISArticle.load(jobdir)
-
-    df3 = spider.scrape(1).get()
-    assert len(df3) >= len(df2) + 1
-
-
-def test_ISArticle_no_params():
-    # Test scraping
-    spider = ISArticle().scrape(10)
-    df = spider.get()
-    assert len(df) == 10
-    assert len(df.columns) == 8
-
-    # Test continuing scraping (poor results, no driver)
-    df2 = spider.scrape(10).get()
-    assert len(df2) >= len(df) + 10
-
-    # Save and load spider
-    jobdir = spider.save()
-    spider = ISArticle.load(jobdir)
-
-    df3 = spider.scrape(10).get()
-    assert len(df3) >= len(df2) + 10
-
-
-def test_ILArticle_with_category():
-    # Test scraping
-    spider = ILArticle('ulkomaat').scrape(1)
-    df = spider.get()
-    assert len(df) >= 1
-    assert len(df.columns) == 8
-
-    # Test continuing scraping
-    df2 = spider.scrape(1).get()
-    assert len(df2) >= len(df) + 1
-
-    # Save and load spider
-    jobdir = spider.save()
-    spider = ILArticle.load(jobdir)
-
-    df3 = spider.scrape(1).get()
-    assert len(df3) >= len(df2) + 1
-
-
-def test_ILArticle_no_params():
-    # Test scraping
-    spider = ILArticle().scrape(10)
-    df = spider.get()
-    assert len(df) >= 10
-    assert len(df.columns) == 8
-
-    # Test continuing scraping  
-    df2 = spider.scrape(10).get()
-    assert len(df2) >= len(df) + 10
-
-    # Save and load spider
-    jobdir = spider.save()
-    spider = ILArticle.load(jobdir)
-
-    df3 = spider.scrape(10).get()
-    assert len(df3) >= len(df2) + 10
 
 
 def test_spider_save_load_with_jobdir():
@@ -174,7 +87,7 @@ def test_spider_logging():
 
 
 def test_spider_progress_bar():
-    # Progress bas true by default
+    # Progress bar true by default
     spider = ILArticle()
     spider.scrape(1)
     assert spider.progress_bar == True
